* Fixed: Possible runtime assert
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9016 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/projectdepsdlg.cpp
@@ -70,7 +70,7 @@ bool ProjectDepsDlg::SaveList()
     Manager::Get()->GetProjectManager()->ClearProjectDependencies(thisprj);
 
     // now set the the new deps
-    for (int i = 0; i < (int)lst->GetCount(); ++i)
+    for (unsigned int i = 0; i < lst->GetCount(); ++i)
     {
         if (!lst->IsChecked(i))
             continue;
